django_smart_intrusion_detection/requirements.txt
Django==5.0.4
django-cors-headers==4.3.1
opencv-python==4.9.0.80
torch==2.2.2
torchvision==0.17.2
onnxruntime-gpu==1.17.1